Un-abstracted default value of `BlockProperty`
