+++ src/main/java/dev/psygamer/wireframe/block/state/property/BlockProperty.java
@@ -1,39 +1,48 @@
 package dev.psygamer.wireframe.block.state.property;
 
 import com.google.common.collect.ImmutableList;
-import dev.psygamer.wireframe.util.helper.IFreezable;
+import dev.psygamer.wireframe.internal.block.state.BlockPropertyWrapper;
 
 import java.util.Objects;
 import java.util.Optional;
 
-public abstract class BlockProperty <T extends Comparable<T>> implements IFreezable {
+public abstract class BlockProperty <T extends Comparable<T>> {
 	
 	protected final String propertyName;
+	protected T defaultValue;
 	
-	public BlockProperty(final String propertyName) {
+	public BlockProperty(final String propertyName, final T defaultValue) {
 		Objects.requireNonNull(propertyName, "The property name may not be null");
 		
 		this.propertyName = propertyName;
+		this.defaultValue = defaultValue;
 	}
 	
 	public static <T extends Comparable<T>> BlockProperty<T> get(final net.minecraft.state.Property<T> internalProperty) {
 		if (internalProperty == null)
 			return null;
 		
-		return null;//new BlockProperty<>(internalProperty);
+		return new BlockPropertyWrapper<>(internalProperty);
 	}
 	
 	public String getPropertyName() {
 		return this.propertyName;
 	}
 	
+	public T getDefaultValue() {
+		return this.defaultValue;
+	}
+	
+	protected void setDefaultValue(final T defaultValue) {
+		if (!getPossibleValues().contains(defaultValue))
+			throw new IllegalArgumentException("Default value '" + defaultValue + "' is not a possible value!");
+		
+		this.defaultValue = defaultValue;
+	}
+	
 	public abstract Optional<T> getValue(final String valueName);
 	
 	public abstract String getValueName(final T value);
-	
-	public abstract T getDefaultValue();
-	
-	protected abstract void setDefaultValue(final T defaultValue);
 	
 	public abstract ImmutableList<T> getPossibleValues();
